+++ slab-alloc/CHANGELOG.md
@@ -1,4 +1,4 @@
-<!-- Copyright 2017 the authors. See the 'Copyright and license' section of the
+<!-- Copyright 2017-2018 the authors. See the 'Copyright and license' section of the
 README.md file at the top-level directory of this repository.
 
 Licensed under the Apache License, Version 2.0 (the LICENSE-APACHE file) or
@@ -15,6 +15,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Added this changelog
+
+### Changed
+- Upgraded to new `ObjectAlloc` and `UntypedObjectAlloc` traits that use
+  `NonNull<u8>` instead of `*mut u8`
+- Switch from `*mut T` to `NonNull<T>` or `Option<NonNull<T>>` for various
+  internal pointers
+- Make `PtrHashMap` support any value type that is `Copy`, not just raw
+  pointers
 
 ### Fixed
 - Fixed a bug that prevented compilation on 32-bit Windows